Finding max and min value in python.
Here are the two question which are given in practice for the Coursera course. If we do small mistake in python programming then our code will not work even indent.
Two question are given in the exercise shown in the screenshots are given below
Question 1. Create a dictionary called d that keeps track of all the characters in the string placement and notes how many times each character was seen. Then, find the key with the lowest value in this dictionary and assign that key to min_value.  
Question 2. Create a dictionary called lett_d that keeps track of all of the characters in the string product and notes how many times each character was seen. Then, find the key with the highest value in this dictionary and assign that key to max_value.